# optimizer: refactor SROA pass