- use `BitSet` instead of `IdSet{Int}`
- reduce # of dynamic allocations
- separate some computations into individual functions

---

TTFP improvements:
> before (df0080c7830c1e1af577ea1b88b6b64d6f6d1ffa)
```
❯ ./usr/bin/julia -e "@time using Plots; @time plot(rand(10,3));"
  4.588537 seconds (9.34 M allocations: 624.761 MiB, 6.73% gc time, 15.28% compilation time)
  2.875475 seconds (3.84 M allocations: 214.448 MiB, 5.90% gc time, 99.77% compilation time)
```
> after (cef9efa3152a1f69ddc23c026ac5d966cbe055ef)
```
❯ ./usr/bin/julia -e "@time using Plots; @time plot(rand(10,3));"
  4.543190 seconds (9.24 M allocations: 615.898 MiB, 6.70% gc time, 15.24% compilation time)
  2.823435 seconds (3.78 M allocations: 209.089 MiB, 5.43% gc time, 99.76% compilation time)
```