@Keno 2021-11-26
`BitSet` can be dangerous on large functions with high sparsity