@aviatesk 2021-11-26
> `BitSet` can be dangerous on large functions with high sparsity

Aha, I see the point. SROA works on IR after inlining, which can be very large sometimes, and program counters analyzed by it are often very sparse.
I made them `IdSet{Int}` back, also added some note for this choice.